AI: implement these features on both github page and local server version:
make it so that the previous step and next step can be pressed and hold for equivalent of click multiple time. user should also be able to click on the progress bar below the buttons to skip to a certain step. next, in the action preview, make the bar representing probas aligned and same length for easier comparison. next, make sure the progress display while waiting for the simulation is also displayed when using restart sim button.